requirements.txt
openai
foundry-local-sdk
browser-use
open-webui
# Security note: Review these pinned versions regularly and update as needed